Oops, this isn't in Kotlin yet

+++ shared/main/java/net/ccbluex/liquidbounce/ui/client/GuiAntiForge.java
@@ -70,7 +70,7 @@ public class GuiAntiForge extends WrappedGuiScreen {
     @Override
     public void drawScreen(int mouseX, int mouseY, float partialTicks) {
         representedScreen.drawBackground(0);
-        Fonts.fontBold180.drawCenteredString("AntiForge", (int) (representedScreen.getWidth() / 2F), (int) (representedScreen.getHeight() / 8F + 5F), 0xFFFFFFFF.toInt(), true);
+        Fonts.fontBold180.drawCenteredString("AntiForge", (int) (representedScreen.getWidth() / 2F), (int) (representedScreen.getHeight() / 8F + 5F), 0xFFFFFFFF, true);
 
         super.drawScreen(mouseX, mouseY, partialTicks);
     }
